--- BoxOfficeUnitTest/MovieData.json
@@ -1,209 +1,0 @@
-
-{
-  "boxOfficeResult": {
-    "boxofficeType": "일별 박스오피스",
-    "showRange": "20221214~20221214",
-    "dailyBoxOfficeList": [
-      {
-        "rnum": "1",
-        "rank": "1",
-        "rankInten": "0",
-        "rankOldAndNew": "NEW",
-        "movieCd": "20225061",
-        "movieNm": "아바타: 물의 길",
-        "openDt": "2022-12-14",
-        "salesAmt": "4189094718",
-        "salesShare": "85.2",
-        "salesInten": "4189094718",
-        "salesChange": "100",
-        "salesAcc": "4204214718",
-        "audiCnt": "359031",
-        "audiInten": "359031",
-        "audiChange": "100",
-        "audiAcc": "359787",
-        "scrnCnt": "2797",
-        "showCnt": "10115"
-      },
-      {
-        "rnum": "2",
-        "rank": "2",
-        "rankInten": "-1",
-        "rankOldAndNew": "OLD",
-        "movieCd": "20227762",
-        "movieNm": "올빼미",
-        "openDt": "2022-11-23",
-        "salesAmt": "310191475",
-        "salesShare": "6.3",
-        "salesInten": "-315351871",
-        "salesChange": "-50.4",
-        "salesAcc": "26358474567",
-        "audiCnt": "32251",
-        "audiInten": "-33977",
-        "audiChange": "-51.3",
-        "audiAcc": "2684230",
-        "scrnCnt": "765",
-        "showCnt": "1610"
-      },
-      {
-        "rnum": "3",
-        "rank": "3",
-        "rankInten": "-1",
-        "rankOldAndNew": "OLD",
-        "movieCd": "20228313",
-        "movieNm": "오늘 밤, 세계에서 이 사랑이 사라진다 해도",
-        "openDt": "2022-11-30",
-        "salesAmt": "130460332",
-        "salesShare": "2.7",
-        "salesInten": "-59809664",
-        "salesChange": "-31.4",
-        "salesAcc": "2294881377",
-        "audiCnt": "13025",
-        "audiInten": "-6746",
-        "audiChange": "-34.1",
-        "audiAcc": "230017",
-        "scrnCnt": "302",
-        "showCnt": "465"
-      },
-      {
-        "rnum": "4",
-        "rank": "4",
-        "rankInten": "0",
-        "rankOldAndNew": "NEW",
-        "movieCd": "20228640",
-        "movieNm": "신비아파트 극장판 차원도깨비와 7개의 세계",
-        "openDt": "2022-12-14",
-        "salesAmt": "80562429",
-        "salesShare": "1.6",
-        "salesInten": "80562429",
-        "salesChange": "100",
-        "salesAcc": "468198564",
-        "audiCnt": "8166",
-        "audiInten": "8166",
-        "audiChange": "100",
-        "audiAcc": "45836",
-        "scrnCnt": "479",
-        "showCnt": "731"
-      },
-      {
-        "rnum": "5",
-        "rank": "5",
-        "rankInten": "-1",
-        "rankOldAndNew": "OLD",
-        "movieCd": "20210364",
-        "movieNm": "탄생",
-        "openDt": "2022-11-30",
-        "salesAmt": "47009299",
-        "salesShare": "1.0",
-        "salesInten": "-55718345",
-        "salesChange": "-54.2",
-        "salesAcc": "2422899354",
-        "audiCnt": "5297",
-        "audiInten": "-6347",
-        "audiChange": "-54.5",
-        "audiAcc": "270986",
-        "scrnCnt": "209",
-        "showCnt": "265"
-      },
-      {
-        "rnum": "6",
-        "rank": "6",
-        "rankInten": "-3",
-        "rankOldAndNew": "OLD",
-        "movieCd": "20205106",
-        "movieNm": "압꾸정",
-        "openDt": "2022-11-30",
-        "salesAmt": "34829197",
-        "salesShare": "0.7",
-        "salesInten": "-114486570",
-        "salesChange": "-76.7",
-        "salesAcc": "5584017192",
-        "audiCnt": "4535",
-        "audiInten": "-12515",
-        "audiChange": "-73.4",
-        "audiAcc": "590504",
-        "scrnCnt": "161",
-        "showCnt": "186"
-      },
-      {
-        "rnum": "7",
-        "rank": "7",
-        "rankInten": "0",
-        "rankOldAndNew": "NEW",
-        "movieCd": "20183772",
-        "movieNm": "태일이",
-        "openDt": "2021-12-01",
-        "salesAmt": "18730000",
-        "salesShare": "0.4",
-        "salesInten": "18730000",
-        "salesChange": "100",
-        "salesAcc": "1009735590",
-        "audiCnt": "1873",
-        "audiInten": "1873",
-        "audiChange": "100",
-        "audiAcc": "121434",
-        "scrnCnt": "17",
-        "showCnt": "17"
-      },
-      {
-        "rnum": "8",
-        "rank": "8",
-        "rankInten": "-3",
-        "rankOldAndNew": "OLD",
-        "movieCd": "20227885",
-        "movieNm": "프레이 포 더 데블",
-        "openDt": "2022-12-07",
-        "salesAmt": "12480500",
-        "salesShare": "0.3",
-        "salesInten": "-41026688",
-        "salesChange": "-76.7",
-        "salesAcc": "653116987",
-        "audiCnt": "1613",
-        "audiInten": "-4521",
-        "audiChange": "-73.7",
-        "audiAcc": "64413",
-        "scrnCnt": "52",
-        "showCnt": "57"
-      },
-      {
-        "rnum": "9",
-        "rank": "9",
-        "rankInten": "-1",
-        "rankOldAndNew": "OLD",
-        "movieCd": "20228655",
-        "movieNm": "스페이스 키드: 우주에서 살아남기",
-        "openDt": "2022-12-07",
-        "salesAmt": "8707000",
-        "salesShare": "0.2",
-        "salesInten": "-12996500",
-        "salesChange": "-59.9",
-        "salesAcc": "186374130",
-        "audiCnt": "1321",
-        "audiInten": "-1678",
-        "audiChange": "-56",
-        "audiAcc": "22763",
-        "scrnCnt": "35",
-        "showCnt": "37"
-      },
-      {
-        "rnum": "10",
-        "rank": "10",
-        "rankInten": "2",
-        "rankOldAndNew": "OLD",
-        "movieCd": "20228308",
-        "movieNm": "극장판 뽀로로와 친구들: 바이러스를 없애줘!",
-        "openDt": "2022-12-01",
-        "salesAmt": "7574700",
-        "salesShare": "0.2",
-        "salesInten": "-11409100",
-        "salesChange": "-60.1",
-        "salesAcc": "1254979300",
-        "audiCnt": "961",
-        "audiInten": "-1398",
-        "audiChange": "-59.3",
-        "audiAcc": "139723",
-        "scrnCnt": "120",
-        "showCnt": "132"
-      }
-    ]
-  }
-}
